show user info

+++ JLOSChina/Models/OSCUserActiveTimelineModel.h
@@ -11,7 +11,7 @@
 
 @interface OSCUserActiveTimelineModel : OSCBaseTableModel
 
-@property (nonatomic, assign) unsigned long userId;
+@property (nonatomic, assign) long long userId;
 @property (nonatomic, copy) NSString* username;
 @property (nonatomic, strong) OSCUserFullEntity* userEntity;
